requirements.txt
arrow==0.17.0
asgiref==3.3.0
beautifulsoup4==4.9.3
blessed==1.17.11
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
django-picklefield==3.0.1
django-q==1.3.4
djangorestframework==3.12.2
idna==2.10
mysqlclient==2.0.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
wcwidth==0.2.5
